Move codefix types into services/types.ts
